refac(calcdex): reorganized components
