+++ src/pages/Calcdex/CalcdexError/CalcdexError.module.scss
