+++ src/pages/Calcdex/FieldCalc/FieldCalc.tsx
@@ -4,25 +4,21 @@ import { Dropdown } from '@showdex/components/form';
 import { TableGrid, TableGridItem } from '@showdex/components/layout';
 import { ToggleButton } from '@showdex/components/ui';
 import {
-  // LegacyWeatherNames,
   PlayerSideConditionsDexMap,
   PlayerSideConditionsToggleMap,
   PlayerSideScreensToggleMap,
   TerrainDescriptions,
   TerrainNames,
   WeatherDescriptions,
-  // WeatherMap,
-  // WeatherNames,
 } from '@showdex/consts/field';
 import { useColorScheme } from '@showdex/redux/store';
 import { formatId } from '@showdex/utils/app';
 import { getDexForFormat, getWeatherConditions } from '@showdex/utils/dex';
-// import type { GenerationNum } from '@smogon/calc';
 import type { Weather } from '@smogon/calc/dist/data/interface';
 import type { DropdownOption } from '@showdex/components/form';
 import type { CalcdexBattleField, CalcdexPlayerKey } from '@showdex/redux/store';
 import type { ElementSizeLabel } from '@showdex/utils/hooks';
-import { useCalcdexContext } from './CalcdexProvider';
+import { useCalcdexContext } from '../CalcdexContext';
 import styles from './FieldCalc.module.scss';
 
 interface FieldCalcProps {
