+++ src/pages/Calcdex/PokeCalc/index.ts
@@ -1,0 +1,1 @@
+export * from './PokeCalc';
