+++ src/pages/Calcdex/PokeInfo/index.ts
@@ -1,0 +1,1 @@
+export * from './PokeInfo';
