+++ src/pages/Calcdex/PokeMoves/index.ts
@@ -1,0 +1,1 @@
+export * from './PokeMoves';
